feat(item definitions): display name must now be implemented on all ItemDefinitionBase classes
For something like a card with a lot of file logic behind it, the name of the card can be safely
hard coded. ItemDefinitionBase no longer contains a serialized implementation of string _displayName
so you'll have to implement it

BREAKING CHANGE: Find all classes that inherit ItemDefinitionBase and add `[SerializeField] string
_displayName; public override string DisplayName => _displayName;`. To fix the error that your
classes do not implement  the display name get accessor.